+++ src/dreamstate/core/internals/internals.test.ts
@@ -9,7 +9,7 @@ import {
   CONTEXT_SIGNAL_HANDLERS_REGISTRY,
   CONTEXT_SIGNAL_METADATA_REGISTRY,
   CONTEXT_QUERY_METADATA_REGISTRY,
-  SIGNAL_LISTENERS_REGISTRY
+  SIGNAL_LISTENERS_REGISTRY, CONTEXT_SERVICES_REFERENCES
 } from "@/dreamstate/core/internals/index";
 import { unRegisterService } from "@/dreamstate/core/registry/unRegisterService";
 import { registerService } from "@/dreamstate/test-utils/registry/registerService";
@@ -20,7 +20,7 @@ describe("Dreamstate internals", () => {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     const internals = require("./internals");
 
-    expect(Object.keys(internals)).toHaveLength(11);
+    expect(Object.keys(internals)).toHaveLength(12);
   });
 
   it("Internals should be initialized", () => {
@@ -34,6 +34,7 @@ describe("Dreamstate internals", () => {
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(0);
 
     expect(CONTEXT_SERVICES_REGISTRY).toBeInstanceOf(WeakMap);
+    expect(CONTEXT_SERVICES_REFERENCES).toBeInstanceOf(WeakMap);
     expect(CONTEXT_OBSERVERS_REGISTRY).toBeInstanceOf(WeakMap);
     expect(CONTEXT_QUERY_METADATA_REGISTRY).toBeInstanceOf(WeakMap);
     expect(CONTEXT_SIGNAL_HANDLERS_REGISTRY).toBeInstanceOf(WeakMap);
@@ -51,6 +52,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_SIGNAL_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_SIGNAL_HANDLERS_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
+    expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBeUndefined();
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(0);
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();
@@ -64,6 +66,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_SIGNAL_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined(); // No signal listeners here.
     expect(CONTEXT_SIGNAL_HANDLERS_REGISTRY.get(TestContextManager)).toBeDefined();
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined(); // No query listeners here.
+    expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBe(0);
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(1);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(1);
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();
@@ -76,6 +79,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_SUBSCRIBERS_REGISTRY.get(TestContextManager)).toBeDefined();
     expect(CONTEXT_SIGNAL_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
+    expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBe(0);
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(0);
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();