+++ src/dreamstate/core/internals/internals.ts
@@ -7,7 +7,8 @@ import type {
   TSignalSubscriptionMetadata,
   TUpdateObserver,
   TUpdateSubscriber,
-  TDreamstateService
+  TDreamstateService,
+  TAnyObject
 } from "@/dreamstate/types";
 
 /**
@@ -23,6 +24,9 @@ export const EMPTY_ARR: [] = [];
 // Currently active services instance types.
 export const CONTEXT_SERVICES_ACTIVATED: Set<TDreamstateService> = new Set();
 
+// Currently active services instance types.
+export const CONTEXT_SERVICES_REFERENCES: TServiceMap<number> = new WeakMap();
+
 // ReactContext registry, lazy initialized constants.
 export const CONTEXT_REACT_CONTEXTS_REGISTRY: TServiceMap<Context<any>> = new WeakMap();
 
@@ -30,7 +34,7 @@ export const CONTEXT_REACT_CONTEXTS_REGISTRY: TServiceMap<Context<any>> = new We
 export const CONTEXT_SERVICES_REGISTRY: TServiceMap<InstanceType<TDreamstateService>> = new WeakMap();
 
 // ContextManagers context references registry.
-export const CONTEXT_STATES_REGISTRY: TServiceMap<object> = new WeakMap();
+export const CONTEXT_STATES_REGISTRY: TServiceMap<TAnyObject> = new WeakMap();
 
 // ContextManagers observers registry - Providers of related context.
 export const CONTEXT_OBSERVERS_REGISTRY: TServiceMap<Set<TUpdateObserver>> = new WeakMap();