+++ src/dreamstate/core/observing/createManagersObserver.ts
@@ -1,6 +1,19 @@
-import { ComponentType, createElement, memo, ReactElement, useCallback, useEffect, useMemo, useState } from "react";
+import {
+  ComponentType,
+  createElement,
+  memo,
+  MutableRefObject,
+  ReactElement,
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState
+} from "react";
 
 import { EMPTY_ARR } from "@/dreamstate/core/internals";
+import { startServiceObserving } from "@/dreamstate/core/observing/startServiceObserving";
+import { stopServiceObserving } from "@/dreamstate/core/observing/stopServiceObserving";
 import { provideSubTreeRecursive } from "@/dreamstate/core/provision/provideSubTreeRecursive";
 import { addServiceObserverToRegistry } from "@/dreamstate/core/registry/addServiceObserverToRegistry";
 import { registerService } from "@/dreamstate/core/registry/registerService";
@@ -15,11 +28,10 @@ import { IStringIndexed, TAnyContextManagerConstructor, TDreamstateService } fro
 export function createManagersObserver(children: ComponentType | null, sources: Array<TDreamstateService>) {
   if (!Array.isArray(sources)) {
     throw new TypeError(
-      "Wrong provider dreamstate supplied. Only array of context services is acceptable."
+      "Wrong providers parameter supplied. Only array of context services is acceptable."
     );
   }
 
-  // Validate sources.
   for (let it = 0; it < sources.length; it ++) {
     if (!sources[it] || !(sources[it].prototype instanceof ContextService)) {
       throw new TypeError("Only classes extending ContextService can be supplied for provision.");
@@ -28,37 +40,114 @@ export function createManagersObserver(children: ComponentType | null, sources: 
 
   // todo: Validate duplicates for dev bundle? Should not be an issue since context value is always same.
 
-  // Check only managers with required provision.
+  /**
+   * Check only managers with required provision.
+   * Do not include services for subTree rendering but add registering logic for services.
+   */
   const managers: Array<TAnyContextManagerConstructor> = sources.filter(function(Service: TDreamstateService) {
     return Service.prototype instanceof ContextManager;
   }) as Array<TAnyContextManagerConstructor>;
 
-  // Create observer component that will handle observing.
+  /**
+   * Create observer component that will handle observing.
+   */
   function Observer(props: IStringIndexed<any>): ReactElement {
-    // Update providers subtree utility.
-    const [ , updateState ] = useState();
+    const [ , forceRender ] = useState({});
     const updateProviders = useCallback(function() {
-      updateState({});
+      forceRender({});
     }, EMPTY_ARR);
 
-    // Subscribe to tree updater and lazily get first context value.
-    for (let it = 0; it < sources.length; it ++) {
-      useMemo(function(): void {
+    /**
+     * Handle internal observing flags shared between re-renders and lifecycle effects.
+     */
+    const viewState: MutableRefObject<{
+      nextObservedSources: Array<TDreamstateService>;
+      observedSources: Array<TDreamstateService>;
+      isInitialProvision: boolean;
+      isProvisionDisposing: boolean;
+    }> = useRef({
+      nextObservedSources: sources,
+      observedSources: sources,
+      isInitialProvision: true,
+      isProvisionDisposing: false
+    });
+
+    /**
+     * Use memo for first and single init of required components.
+     * useLayoutEffect will not work for some environments and SSR.
+     *
+     * Note: Shared between components that do mount-unmount is the same node.
+     */
+    useMemo(function(): void {
+      viewState.current.nextObservedSources = sources;
+
+      for (let it = 0; it < sources.length; it ++) {
         registerService(sources[it]);
-      }, EMPTY_ARR);
+      }
+    }, sources);
 
-      useEffect(function() {
-        addServiceObserverToRegistry(sources[it], updateProviders);
+    /**
+     * Mount current observers.
+     * Count references of providers to detect whether we start provisioning or ending it.
+     */
+    useEffect(function() {
+      for (let it = 0; it < viewState.current.observedSources.length; it ++) {
+        addServiceObserverToRegistry(viewState.current.observedSources[it], updateProviders);
+        registerService(viewState.current.observedSources[it]);
+        startServiceObserving(viewState.current.observedSources[it]);
+      }
 
-        /**
-         * Destructor-like order for services unregistering.
-         * React calls all hooks in 0...n order, but it should be like 0...n for start and n...0 for end.
-         */
-        return function() {
-          return removeServiceObserverFromRegistry(sources[sources.length - it - 1], updateProviders);
-        };
-      }, EMPTY_ARR);
-    }
+      /**
+       * Unmount current observers.
+       */
+      return function() {
+        viewState.current.isProvisionDisposing = true;
+
+        for (let it = viewState.current.observedSources.length - 1; it >= 0; it --) {
+          removeServiceObserverFromRegistry(viewState.current.observedSources[it], updateProviders);
+          stopServiceObserving(viewState.current.observedSources[it]);
+        }
+      };
+    }, EMPTY_ARR);
+
+    /**
+     * Update current observers.
+     * Detect whether observers were moved/updated/hot module replacement was activated.
+     */
+    useEffect(function() {
+      if (viewState.current.isInitialProvision) {
+        viewState.current.isInitialProvision = false;
+      } else {
+        for (let it = 0; it < sources.length; it ++) {
+          if (viewState.current.observedSources[it] !== sources[it]) {
+            addServiceObserverToRegistry(sources[it], updateProviders);
+            registerService(sources[it]);
+            startServiceObserving(sources[it]);
+          }
+        }
+      }
+
+      /**
+       * Remember current observed sources after HMR replacement or update of dependencies.
+       */
+      viewState.current.observedSources = sources;
+
+      /**
+       * Clean up previous sources after dependencies updates or HMR.
+       */
+      return function() {
+        if (viewState.current.isProvisionDisposing) {
+          return;
+        } else {
+          for (let it = sources.length - 1; it >= 0; it --) {
+            if (viewState.current.nextObservedSources[it] !== sources[it]) {
+              removeServiceObserverFromRegistry(sources[it], updateProviders);
+              stopServiceObserving(sources[it]);
+            }
+          }
+        }
+      };
+    }, sources);
 
     return provideSubTreeRecursive(children ? createElement(children, props) : props.children, managers, 0);
   }