+++ src/dreamstate/core/registry/addServiceObserverToRegistry.ts
@@ -1,5 +1,4 @@
-import { CONTEXT_SERVICES_REGISTRY, CONTEXT_OBSERVERS_REGISTRY } from "@/dreamstate/core/internals";
-import { ContextService } from "@/dreamstate/core/services/ContextService";
+import { CONTEXT_OBSERVERS_REGISTRY } from "@/dreamstate/core/internals";
 import { TDreamstateService, TUpdateObserver } from "@/dreamstate/types";
 
 /**
@@ -10,9 +9,4 @@ export function addServiceObserverToRegistry(
   observer: TUpdateObserver
 ): void {
   CONTEXT_OBSERVERS_REGISTRY.get(Service)!.add(observer);
-
-  // Notify about provision, if it is first observer.
-  if (CONTEXT_OBSERVERS_REGISTRY.get(Service)!.size === 1) {
-    (CONTEXT_SERVICES_REGISTRY.get(Service) as ContextService)["onProvisionStarted"]();
-  }
 }