+++ src/dreamstate/core/registry/getReactContext.test.ts
@@ -1,13 +1,14 @@
 import { Context } from "react";
 
 import { CONTEXT_REACT_CONTEXTS_REGISTRY } from "@/dreamstate/core/internals";
+import { TAnyObject } from "@/dreamstate/types";
 import { TestContextManager } from "@/fixtures";
 
 describe("getReactContext method", () => {
   it("Related react context should be lazily initialized correctly with changed displayName", () => {
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();
 
-    const contextType: Context<object> = TestContextManager.REACT_CONTEXT;
+    const contextType: Context<TAnyObject> = TestContextManager.REACT_CONTEXT;
 
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeDefined();