+++ src/dreamstate/core/registry/registerService.ts
@@ -4,18 +4,18 @@ import {
   CONTEXT_SIGNAL_HANDLERS_REGISTRY,
   CONTEXT_STATES_REGISTRY,
   CONTEXT_SUBSCRIBERS_REGISTRY,
-  CONTEXT_SERVICES_ACTIVATED
+  CONTEXT_SERVICES_ACTIVATED, CONTEXT_SERVICES_REFERENCES
 } from "@/dreamstate/core/internals";
 import { processComputed } from "@/dreamstate/core/observing/processComputed";
 import { ContextManager } from "@/dreamstate/core/services/ContextManager";
 import { onMetadataSignalListenerCalled } from "@/dreamstate/core/signals/onMetadataSignalListenerCalled";
 import { subscribeToSignals } from "@/dreamstate/core/signals/subscribeToSignals";
-import { TDreamstateService, TSignalListener } from "@/dreamstate/types";
+import { TAnyObject, TDreamstateService, TSignalListener } from "@/dreamstate/types";
 
 /**
  * Register context manager entry.
  */
-export function registerService<T extends object>(
+export function registerService<T extends TAnyObject>(
   Service: TDreamstateService
 ): void {
   // Only if registry is empty -> create new instance, remember its context and save it to registry.
@@ -26,6 +26,7 @@ export function registerService<T extends object>(
     CONTEXT_OBSERVERS_REGISTRY.set(Service, CONTEXT_OBSERVERS_REGISTRY.get(Service) || new Set());
     CONTEXT_SERVICES_REGISTRY.set(Service, instance);
     CONTEXT_SIGNAL_HANDLERS_REGISTRY.set(Service, signalHandler);
+    CONTEXT_SERVICES_REFERENCES.set(Service, 0);
 
     subscribeToSignals(signalHandler);
 
@@ -33,7 +34,7 @@ export function registerService<T extends object>(
 
     // Currently only context managers require additional information initialization.
     if (Service.prototype instanceof ContextManager) {
-      CONTEXT_STATES_REGISTRY.set(Service, (instance as ContextManager<object>).context);
+      CONTEXT_STATES_REGISTRY.set(Service, (instance as ContextManager<TAnyObject>).context);
       // Subscribers are not always sync, should not block their un-sub later.
       CONTEXT_SUBSCRIBERS_REGISTRY.set(
         Service,