+++ src/dreamstate/core/registry/unSubscribeFromManager.ts
@@ -1,11 +1,11 @@
 import { CONTEXT_SUBSCRIBERS_REGISTRY } from "@/dreamstate/core/internals";
 import { ContextManager } from "@/dreamstate/core/services/ContextManager";
-import { IContextManagerConstructor, TUpdateSubscriber } from "@/dreamstate/types";
+import { IContextManagerConstructor, TAnyObject, TUpdateSubscriber } from "@/dreamstate/types";
 
 /**
  * Unsubscribe from manager updates.
  */
-export function unsubscribeFromManager<T extends object, D extends IContextManagerConstructor<T>>(
+export function unsubscribeFromManager<T extends TAnyObject, D extends IContextManagerConstructor<T>>(
   Manager: D,
   subscriber: TUpdateSubscriber<T>
 ): void {