+++ src/dreamstate/core/signals/emitSignal.ts
@@ -9,7 +9,7 @@ import { ISignal, ISignalEvent, TDreamstateService, TSignalListener, TSignalType
 export function emitSignal<D = undefined, T extends TSignalType = TSignalType>(
   base: ISignal<D, T>,
   emitter: TDreamstateService | null = null
-): void {
+): Promise<void> {
   if (!base || base.type === undefined) {
     throw new TypeError("Signal must be an object with declared type.");
   }
@@ -20,12 +20,26 @@ export function emitSignal<D = undefined, T extends TSignalType = TSignalType>(
     cancel: cancelSignal
   });
 
-  // Async processing of subscribed metadata.
-  SIGNAL_LISTENERS_REGISTRY.forEach(function(it: TSignalListener<D, T>) {
-    setTimeout(function() {
-      if (!signal.canceled) {
-        it(signal);
-      }
+  // Async processing of subscribed metadata to prevent exception blocking.
+  // todo: Add test cases for awaiting?
+  return new Promise<void>(function(resolve: () => void): void {
+    const handlersToProcessCount: number = SIGNAL_LISTENERS_REGISTRY.size;
+    let processedHandlers: number = 0;
+
+    SIGNAL_LISTENERS_REGISTRY.forEach(function(it: TSignalListener<D, T>) {
+      setTimeout(function() {
+        try {
+          processedHandlers ++;
+
+          if (!signal.canceled) {
+            it(signal);
+          }
+        } finally {
+          if (processedHandlers === handlersToProcessCount) {
+            resolve();
+          }
+        }
+      });
     });
   });
 }