+++ src/fixtures/signals/EmittingContextManager.ts
@@ -1,9 +1,10 @@
 import { ContextManager } from "@/dreamstate/core/services/ContextManager";
+import { TAnyObject } from "@/dreamstate/types";
 import { ESignal } from "@/fixtures/signals/ESignal";
 
-export class EmittingContextManager extends ContextManager<object> {
+export class EmittingContextManager extends ContextManager<TAnyObject> {
 
-  public context: object = {};
+  public context: TAnyObject = {};
 
   public sendNumberSignal(): void {
     this.emitSignal({ type: ESignal.NUMBER_SIGNAL, data: Math.random() });