+++ src/fixtures/signals/SubscribedFunctionalComponent.ts
@@ -1,4 +1,4 @@
-import { createElement, useState } from "react";
+import { createElement, ReactElement, useState } from "react";
 
 import { useSignals } from "@/dreamstate";
 import { ESignal } from "@/fixtures/signals/ESignal";
@@ -6,7 +6,7 @@ import { TStringSignalEvent } from "@/fixtures/signals/types";
 
 export function SubscribedFunctionalComponent({
   onInternalSignal
-}: { onInternalSignal: (signal: TStringSignalEvent) => void }) {
+}: { onInternalSignal: (signal: TStringSignalEvent) => void }): ReactElement {
   const [ value, setValue ] = useState("initial");
 
   useSignals((signal: TStringSignalEvent) => {